Pg updates

+++ db/migrate/20130905101725_alter_specifications.rb
@@ -1,7 +1,7 @@
 class AlterSpecifications < ActiveRecord::Migration
   def change
-    	add_column :quality_specifications, :cvi, :double
-    	add_column :quality_specifications, :cvw, :double
+    	add_column :quality_specifications, :cvi, :decimal
+    	add_column :quality_specifications, :cvw, :decimal
     	add_column :quality_specifications, :goaltype, :string
     	add_column :test_codes,:readcode, :string
     	remove_column :test_codes, :testCode
