+++ db/migrate/20130905103936_alter_specification_add_specification_te.rb
@@ -1,7 +1,7 @@
 class AlterSpecificationAddSpecificationTe < ActiveRecord::Migration
   def change
-    add_column :quality_specifications, :te, :double
-    add_column :quality_specifications, :cv_d, :double
-    add_column :quality_specifications, :bias_d, :double
+    add_column :quality_specifications, :te, :decimal
+    add_column :quality_specifications, :cv_d, :decimal
+    add_column :quality_specifications, :bias_d, :decimal
   end
 end
